+++ changelog.md
@@ -8,9 +8,14 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
 ### Changed
 - `strengthen` no longer returns an `Either`, since the proof
   that it should always succeed is in its constraints.
+- `validate` now takes a `Proxy` as its first argument.
 
 ### Removed
 - Refined.These module
+
+### Fixed
+- bug in `sized` internal helper that caused formatting issues
+  in sized predicate errors
 
 ## [0.5.1] - 2020-07-14
 ### Changed